Completed code for MWE_InternalRTCBlink
Used an alarm generated with the internal RTC to blink an LED. Have to set the alarm time to the target time less 1 second, because for some reason when the current time is retrieved with the API, the seconds are rounded up.

Verified that the code works when seconds overflows.